[roll third_party/pigweed] pw_assert_basic: Fix build and minor bugfix
This commit fixes the bazel build files. This fix also addresses a bug
where an interaction between @io_bazel//src/main/tools:process-wrapper
and pw_assert_basic caused the output from `bazel test
//some_test_target` to be completely empty on assertion. This occured
because pw_assert_basic was 'aborting' causing the core to dump without
calling the destructors. Under normal circumstances this is fine as
stdout is synchronised. However as bazel's process wrapper binds
stdout, it results in stdout buffering, meaning that when
pw_assert_basic 'aborts' the buffered stdout is lost. The replacement
of 'abort' to 'exit' allows destructors for the stdout handles to be
called thus flushing the buffer out to bazels process-wrapper.
This is not a critical bug and is more of a nuisance as the message that
prints the line number that caused the assertion is lost when the
program aborts.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/34621

third_party/pigweed Rolled-Commits: 975a3c26b9abf2c..be3fba1ec8d99d0
Roller-URL: https://ci.chromium.org/b/8853211718726833632
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8853211718726833632
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I494af5068dff164b742334f28e6a8d665eb9e5ce
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/36503
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>