[stm32] Add new target for stm32f429i-disc1 w/ HAL
This is a new target for the discovery board based on the in-tree
target but with stm32cubef4 HAL support.

The main differences are that it uses STM's code for initializing the
FPU, reclocks the board to its 180Mhz maximum, and uses the stm32cubef4
backends for blinky.

It does not yet support the test runner, but the existing one in-tree
should work with all STM32F4xx boards that have a ST-Link/v2.1 with
minimal changes.

Change-Id: I2bdd17d5958b5204bf9d2b44909bd8bd1e25402c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/39770
Commit-Queue: Varun Sharma <vars@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>