[roll third_party/pigweed] pw_stream: Allow staging data to MemoryWriter end
Changes the memory writer to use std::memmove to allow data to be staged
to the unused portion of a MemoryWriter's buffer before it is written.
This is useful for operations that require direct access to a buffer to
build data before it is flushed to a MemoryWriter.

No-Docs-Update-Reason: Optimization that shouldn't be advertised
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/40244

third_party/pigweed Rolled-Commits: 9a53d6c97f65375..f82d95b680c9b19
Roller-URL: https://ci.chromium.org/b/8848116767675177168
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8848116767675177168
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I8585f5540215ddb64a9666e6787a29a9cc4c4f9d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/43780
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>