[roll third_party/pigweed] pw_env_setup: Fix pip upgrading on Mac
When bootstrapping with the system python on Mac, __PYVENV_LAUNCHER__
is set in the environment which causes pip installs in the virtualenv
to not work properly. This was fixed for ninja installs but not for
other calls to pip such as the initial upgrade of pip itself.

Fix this by always running pip in the virtualenv during bootstrap.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/58600

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 00ba31c51c7a90e..8552c61aaccdd57
Roller-URL: https://ci.chromium.org/b/8837733030737278417
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8837733030737278417
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I7fd1785821a6380b108e5b2cd5b7e09a0a1523f7
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/58641
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>