[roll third_party/pigweed] pw_rpc: Update C++ client API
This updates the generated service client API to work with an RPC client
and channel ID instead of a channel itself. This will allow removing the
client reference from the channel object.

As part of this, the generated client classes are made instantiable to
improve usability. A client and channel ID must be passed in when
creating a service client; beyond this, they are not needed to invoke
methods.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/60061

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: f345faa70680dfc..73687fb473489a2
Roller-URL: https://ci.chromium.org/b/8836156642737955633
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8836156642737955633
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ie48e925bd09eb7fa01b3260b42d7942fa1327c3d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/60697
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>