[roll third_party/pigweed] pw_software_update: Add mechanism to enable/disable bundle transfer
Adds a mechanism to the BundledUpdateBackend to enable the transfer
handler and return the transfer_id as part of the update start.

In addition a method is added to disable the transfer handler as
part of the start of verification or an update abort.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/61799

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: c2b9379cc32b617..f087b00fa2d156d
Roller-URL: https://ci.chromium.org/b/8835324695860972945
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8835324695860972945
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I1dafbee284eae925c1893c229f3222abd8956aba
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/61835
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>