[roll third_party/pigweed] clang-tidy: readability-inconsistent-declaration-parameter-name
This check verifies that the parameter names used in the function
declarations and definitions match. The fix involves changing either the
declaration or the definition; I chose to do one or the other on a case
by case basis.  My guiding principles were (1) use the same names for
the same parameter in different functions in the same header, and (2)
generally prefer to retain the name in the header over that in the
source file (since it's more public), especially if that name also
appears in a documenting comment on the function.

Original-Bug: 45
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/66640

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 247595e43680a33..4180d4de9a9cd5c
Roller-URL: https://ci.chromium.org/b/8832089975538047025
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8832089975538047025
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5f0f4f980cddf18d9ff212ffec6f0ff2b13eec5c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/66781
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>