[roll third_party/pigweed] pw_software_update/py: Relax root metadata signing checks
When signing a root metadata, the signing key does not need to be listed
in the root metadata. This CL removes those checks.

Use case: When rotating a root key, the new root metadata need to signed
with both the existing root key and the new root key but the old root
key does NOT need to be listed in the new root metadata. In fact if the
rotation aims to revoke the old root key, the old root key MUST NOT be
listed in the new root signature requirements.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/69360

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: e2e23280635baad..4a555a4db0f659b
Roller-URL: https://ci.chromium.org/b/8829157559878112625
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8829157559878112625
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ia28336eb3b5cda148688e0babecc9f2f68fec02e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/71321
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>