[roll third_party/pigweed] pw_presubmit: Fail gracefully if missing base rev
Fail gracefully if 'pw presubmit --base HEAD~1' is run but there is no
'HEAD~1' revision. This happens when dealing with a brand new repository
or when a checkout has a depth of 1. If we can't process a base commit
we print a warning and resume as if the '--base' argument was not used.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/71781

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: e23423ef3e3acef..acc897337259522
Roller-URL: https://ci.chromium.org/b/8828847577568110369
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8828847577568110369
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Icb601e415d47c500baea94ce118b0525d6c7a741
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/71821
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>