[roll third_party/pigweed] pw_software_update: Mandate backend VerifyTargetFile()
VerifyTargetFile() is called for any target files referenced in the
Targets metadata that is not found in the incoming bundle -- e.g.
when some target files have been "personalized out".

The backend MUST implement this function to uphold the security
statement that says "the incoming bundle is a complete and officially
endorsed software update". Without this check, a cracker could serve
an incomplete update -- e.g. by removing some important target
payloads that are supposed to be included.

The actual implementation typically involves retrieving the original
payload of @target_file_name, measuring / hashing it, and comparing
that measurement (hash digest) against records in the (now trusted)
@manifest. The on-device measurement may be cached (preferrably in a
tamper resistant storage) to improve performance on subsequent
verifications.

If the target payload can't not be retrieved, an alternative
verification mechanism could be used such as retrieving the version
number via a target specific safe protocol and comparing that to what
is recorded in @manifest. This is less reliable than hashing and is
a product decision.

In the absense of a good verification mechanism, the target element
SHALL ALWAYS be included in the incoming bundle.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/87320

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 28792b3c57e5410..b58e3aa0731e293
Roller-URL: https://ci.chromium.org/b/8820218339889624657
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8820218339889624657
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Iadf5767dce8bb8c5056777c83931a04e9f3b7830
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/87400
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>