[roll third_party/pigweed] pw_protobuf: StreamEncoder and MemoryEncoder support for tables
This implements protobuf encoding from a struct described by a series of
MessageField structures.

A new protected StreamEncoder::Write() method is added that accepts a
span of bytes corresponding to a 'struct Message', and a span of
MessageField describing the fields therein.

The generated code is extended to add a Write() method to each
StreamEncoder and MemoryEncoder wrapper that accepts a struct of the
correct type and passes in the corresponding fields table.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/89641

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 1d3fe223eee7b73..3c4cb3028073620
Roller-URL: https://ci.chromium.org/b/8817407608285964417
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8817407608285964417
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ie605d1a5d46da1608c53f4b166b43c1eee7eaf6c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/90315
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>