[roll third_party/pigweed] pw_i2c: Add ProbeTransaction to test probes with Mock Initiator
Because Probe is implemented as a one-byte read, Probes must be
accounted for as ReadTransactions when testing with the Mock Initiator,
and must be constructed with a one-byte ignored_buffer with arbitrary
contents.

To clean up testing that might include Probes, add
ProbeTransaction to the Mock Initiator. This constructs the one-byte
Read transaction with an ignored buffer under the hood, but abstracts
this detail away from the caller. This does not change the one-byte read
behavior of Probe.

- Add unit test for Probe in Mock Initiator
- Add Probe Transaction to Mock Initiator header
- Add Probe example to docs

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/90920

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 3e76583a953125f..19b6cf7ca622ef4
Roller-URL: https://ci.chromium.org/b/8816890243369511729
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8816890243369511729
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Id2b265b6e6e0b2f67546655518a706c8407fa68d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/91144
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>