[roll third_party/pigweed] pw_build: Add ability to exec multi argument command
There is an assumption in exec.py that the first
argument will always be exactly 1 command with no
additional arguments.  However for RBE this becomes
a multi argument string and when that multi argument
string is fed into subprocess in exec.py, subprocess
crashes.  This checks to see if there are any spaces
in the command splits the command into a list so
subprocess can handle it correctly whether there is
a space or not.

Test: Verified this generated the build files correct and built
      correctly with RBE and regular builds.
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/98442

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: dd1abf58412617e..fd2a7f66604c3af
Roller-URL: https://ci.chromium.org/b/8807450871352111361
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8807450871352111361
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I29d91a9b9e3f88e0b96ce5b0e524abce003b0208
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/103669
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>