[roll third_party/pigweed] pw_result: Add transform monadic operation
This CL adds a monadic operation to pw::Result<T> for `transform`. This
method takes a function, which will be invoked only if the pw::Result<T>
contains a value. If the pw::Result<T> contains a status, then that
status will be returned. This is useful for applying a function to a
Result and transforming the type.

 Result<int> ConvertStringToInteger(std::string_view);
 int MultiplyByTwo(int x) { return x * 2; }

 Result<int> x = ConvertStringToInteger("42")
                   .transform(MultiplyByTwo);

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/101348

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 9b7743e6e8faa10..b568966d401b1b3
Roller-URL: https://ci.chromium.org/b/8806177475407806257
Cq-Cl-Tag: roller-builder:pigweed-experimental-roller
Cq-Cl-Tag: roller-bid:8806177475407806257
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I3d67ec56ff8b3499b82562c7c454d968589273de
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/105887
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>