[roll third_party/pigweed] pw_build: Support -include args in pw_linker_script()
Add a new "includes" argument to the pw_linker_script() template to
include from the command line a list of files.

There was a missing "invoker." when using cflags that was causing it to
fail if cflags was passed to pw_linker_script(), which is also fixed
here.

Finally, the invoker.inputs list was only being passed to the
pw_source_set created, which would re-link files that depended on this
linker script, but wouldn't re-generate the final linker script by
running the preprocessor, so files listed as "-include" in cflags or
as #include directives would not work. This patch moves the inputs
argument to the pw_exec so changes there also trigger a re-generation of
the final linker script.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/106110

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 6723e0fcaf74e70..3a90f7e868d8ef7
Roller-URL: https://ci.chromium.org/b/8806030560882001377
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I1e615b9ee3ebc9a4fd9f9eef3e58a1f882864387
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/106292
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>