[roll third_party/pigweed] pw_stream: Implement StdFileReader::ConservativeLimit
The ConservativeLimit method returns an estimate of how much data is
available to be read or written in a pw::stream. The default
implementation returns a very large value ("unlimited").

This commit implements a file-specific implementation for StdFileReader.
It determines the size of the file (and thus how much data is left to
read) by seeking to the end and getting the current file position. This
makes the ConservativeReadLimit method much more useful for users of the
class.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/106130

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 26e210529ec4d39..3bac75d77b9b492
Roller-URL: https://ci.chromium.org/b/8805741364250787793
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I05972d4b52d560ccac684cb8f1cc9c296fb135fb
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/106631
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>