[roll third_party/pigweed] pw_build: Don't mutate deps list in bazel macros
The deps list is passed by reference. If the caller uses a variable as
the deps list, rather than a literal, that variable will be mutated as
a side effect of the macro invocation if we use `append`. This is not
intended.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/106912

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: ac7f0090500ed5b..79c1756eeb72e81
Roller-URL: https://ci.chromium.org/b/8805450036365551601
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I49adfbceacf7e4a54684d2bbc800ab495ec69e59
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/107491
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>