[roll third_party/pigweed] pw_protobuf: Distinguish between string and bytes fields
Use a MessageField bit to indicate that the field is a string field, as
opposed to a bytes field. Dedicating a bit to one specific type is an
inefficient use of the field_info_ bits, but there were bits to spare
and this was the simplest way to do this. If more field_info_ bits are
needed in the future, a few of the existing fields could be consolidated
into a single 4-bit field_type that specifies that precise type (e.g.
int32, fixed64, message, etc.).

It is necessary to distinguish between string and bytes in order to
switch from using pw::Vector<char> to pw::InlineString<> for string
fields.

This change also updates C++ codegen to include /*attr_name=*/ comments
for bool arguments to improve readability.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/110231

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: d4260beae8c08c7..364991493b41173
Roller-URL: https://ci.chromium.org/b/8802910154795337953
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Iefb96a34cb71420e5e14704c9d72d1927c3e3b82
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/110523
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>