[roll third_party/pigweed] pw_ide: activate.py
Provides a Python script that can be run independently of the Pigweed
environment, and execute a given shell command *within* an activated
Pigweed environment.

Why do we need this? Editors aren't aware of the Pigweed environment,
and neither are any subprocesses they spawn (unless you launch the
editor *from* an activated environment, which you might do with vim but
probably wouldn't do from VS Code). We wrap tasks that need access to
the Pigweed environment in this script.

Note that this script replicates much of the functionality already
present in the pw_env_setup module. The difference is that this script
can be run outside of the Pigweed environment and has no dependencies.
In the long run, we should find a way to merge the two in a way that
works for both use cases.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/110259

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: aa2b5e3385c4c9a..4769b27af137433
Roller-URL: https://ci.chromium.org/b/8802529912591358849
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ie1db24cb299b6ce0e1aaf0aef3cb7f6b3ed56b71
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/111035
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>