pw_display: Add InitFramebuffer() function.
As not all displays are the same size so move the job of initializing the
framebuffer to pw_display. In this change pw_display::UpdatePixelDouble()
remains, but is unused. It is the responsibility of the display to do
pixel doubling if necessary. pw_display_stm32f429i_disc1_stm32cube_ili9341
has support for pixel doubling (with an #ifdef change), but no other
displays currently support it.

Change-Id: I265ac0bd751faa1120fd6e43d0fd51c0a8553259
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/113270
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>