pw_framebuffer: Remove Pen/Transparent color state
Move the Framebuffer to be a simple owner of the actual framebuffer
pointer along with basic metadata (width, height, etc.). It keeps
basic pixel set/get methods, but drawing state will be managed by
other objects responsible for rendering into the framebuffer. This
means that pw_draw can no longer rely on the fact that the
Framebuffer has a pen color state. This resulted in all pw_draw
functions having an explicit color argument - or more precisely
removing those with no color parameter.

Previously Framebuffer had a concept of a "transparent color"
which was used to indicate an invalid color. Framebuffer::GetPixel()
would return the transparent color if the requested pixel
coordinates were out of bounds. This change modifies GetPixel()
to return a boolean to indicate a valid returned color.

This is to support displays with multiple framebuffers. In this
scenario the drawing state should be maintained elsewhere.
Additionally the Framebuffer object should be small to be
efficiently passed as a stack variable.

Change-Id: I74e5feeac20330d4d07093abb10f6a38df1b3fa3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/114710
Commit-Queue: Chris Mumford <cmumford@google.com>
Reviewed-by: Erik Gilling <konkers@google.com>