Make terminal_demo work with multi-buffer drawing.
Some displays sport multiple framebuffers which means the
terminal demo app can no longer rely on the contents of
the framebuffer to be the last rendered frame. It may be
N-1 (or N-2). This change adds a text buffer where the
last N (currently N=9) lines of pw_log messages are stored.

This lowers the frame rate because the entire display is
redrawn each update cycle. On the STM32F429I-DISC1 the
frame rate went from 37 Hz to 19 Hz.

Change-Id: Idea23fd4720e6a1c5719166a7f0fcfdfa70faaf3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/116490
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>