[roll third_party/pigweed] pw_multisink: Fix drop count after PeekEntry()
When removing messages from a Drain with PeekEntry() and then PopEntry()
of the peeked_entry, it is possible that the multisink receives new
messages between two causing the drain to drop more messages (including
the peeked one). This is not a problem for the peeked entry since the
contents are copied, but on PopEntry() the old message is just ignored
and the drop count not updated.

This patch updates the drain's `last_handled_sequence_id_` in this case
so a subsequent PeekEntry() returns the right number of dropped messages
since the last call to PopEntry().

Test: Extended a unittest to cover this situation.
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/117750

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 8c480d4b331f447..f737777a6bfcafd
Roller-URL: https://ci.chromium.org/b/8798126793742454545
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Id82dc39a2d638ed7310f47cef61f1dc5a50d1a77
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/117850
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>