[roll third_party/pigweed] pw_presubmit: Compose filters in .with_filter()
If .with_filter(f) is called on a Check that already has a filter,
compose the two filters rather than replacing the original filter
completely. Do this by appending each field of the new filter onto
the existing filter (except the name, where the new name is preferred).

In most cases this will make sense, as in the pragma_once check, which
is initially defined with a filter to limit it to C++ header extensions
but excluding files ending with '.pb.h', which can then be composed with
a filter that says to exclude paths beginning with 'third_party/'. (With
the previous behavior the application of the new filter would apply the
Check to all files not under 'third_party/', regardless of extension.)

In some cases this will not make sense, like if the original filter
specified all C++ extensions but a subsequent filter wanted to limit
that to only C++ header files. In that case something other than
.with_filter() will be required.

Also add .unfiltered() which will create a new check with no filters. To
reproduce the previous behavior callers need to use
.unfiltere().with_filter(f).

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/118590

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 47191e27630009f..c86499846c35ec0
Roller-URL: https://ci.chromium.org/b/8797837766733996737
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ib8eafe88d77327db6ac266617ca282b203cb9799
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/118791
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>