pw_display: Moved pico ili9341 to use backend.
The pw::display::backend::Display class was introduced in f7bb4962d5d35
which provides a mechanism for the various display backends to
encapsulate the platform specific implementation.
pw_display_pico_ili9341 was still using the old mechanism - which
was a static class in the anonymous namespace. This change
updates this modules implementation to simplify and allow for
multiple instances.

Change-Id: Ie11e0548991817dd3ba2b5fc9ed4a8ff151c5b18
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/117796
Commit-Queue: Chris Mumford <cmumford@google.com>
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>