[roll third_party/pigweed] pw_sync_freertos: Suspend Scheduler Within ISL
The FreeRTOS backend implements ISL using an interrupt-disabled
critical section. This may be problematic in the event the
holder of the lock calls a kernel API that results in a higher
priority task becoming unblocked, based on the target-specific
implementation of portYIELD.

If the target port implements portYIELD synchronously, the ready
task will immediately preempt the current-running task - despite
being within a critical section.  As part of the context switch,
the interrupts may also be re-enabled.  These side effects
break the expected behavior of the lock.

This condition may not occur in the event portYIELD is implemented
asynchronously however - as the interrupt required to service
the resulting context switch may be masked for the duration of
the critical section.

This CL modifies the ISL implementation of the ISL to additionally
call vTaskSuspendAll()/xTaskResumeAll() on lock/unlock operations.
This fixes the errant condition, as FreeRTOS ensures that the
current-running task will not be preempted while the scheduler
is suspended.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/119810

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: f3eddd6585abdc2..4366c39ac223d3b
Roller-URL: https://ci.chromium.org/b/8795960130675328769
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I20e9dacdceab05ebae330326429bb8d4978395a5
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/122070
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>