Create pw_display_driver_st7789
Move the driver code out of pw_display_pico_st7789_spi and make
it platform agnostic.

This change introduces a new class, pw::display_driver::SPIHelper,
which is a temporary solution to the need to reconfigure the SPI
bus. This was needed because the ST7789 needs to switch between
8 and 16 bit mode for each display update. This will eventually
be removed when implementing a full fix for b/251033990, but this
is sufficient to keep the DisplayDriver platform independent
with minimal hacks.

Bug: none
Change-Id: I85917ac242074581615a0bb5ae260c2dd52a1822
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/119734
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>