display: InitFramebuffer => GetFramebuffer/ReleaseFramebuffer.
Deleted pw::display::InitFramebuffer() and Update() and replaced with
GetFramebuffer() and ReleaseFramebuffer(). This is to support a
future change where the display supports multiple framebuffers
so the application won't be able to hold on to a framebuffer for
the application lifetime. In this new approach the rendering loop
calls GetFramebuffer(), draws to the framebuffer, and then calls
ReleaseFramebuffer() to return it to the display and commence
the pixel transport to the screen.

Change-Id: Id36a783aa967fc73c5c392d60d1792ac255152ce
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/124391
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>