[roll third_party/pigweed] pw_presubmit: Invoke ninja via pw-wrap-ninja, handle SIGTERM
This commit switches the standard pw_presubmit 'ninja' helper to invoke
Ninja via pw-wrap-ninja, and enable the flag to log action start/end
times.

Additionally, it adds (and uses) an option to the 'call' helper to
propagate a SIGTERM to the subprocess. This is helpful because CI may
terminate pw_presubmit with a SIGTERM if it times out, giving it a
chance to do cleanup. When pw-wrap-ninja receives a SIGTERM, it prints
out any actions it was running. Combined, this makes it easy to see
what's running when a Ninja step in a CI job times out.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126936

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 5b53103e1076a70..f19e9283a36eebc
Roller-URL: https://ci.chromium.org/b/8790776023220663697
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ib286ca83e17e55170ff96a9c5696b8c64aa582ce
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/126953
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>