pw_display: Added pw_display_DISPLAY_RESIZE build variable
Added a pw_display_DISPLAY_RESIZE build variable to disable
framebuffer resizing when updating. This defaults to disabled
in the pw_display module as display resizing is currently
only used for development purposes when RAM is tight.
This allows the linker to optimize out UpdateNearestNeighbor
as well as the DisplayDriver::WriteRow() methods.

Change-Id: I8a39a754fbb5f45084dff09df11f471f786a4ac2
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/130331
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>