pw_display: Do not fail if DISPLAY_RESIZE==0 and sizes mismatch
Not all display/framebuffer configurations have equal sizes
and its not always clear when a resize should be done, or
when to pass through to the display driver. For example, with
the NXP round display the following values are used:

  display:
    width: 400
    height: 392
    start.x: 4
  framebuffer:
    width: 392
    height: 392

At present, until a better understanding of display values exist,
allow the display driver to render the framebuffer if DISPLAY_RESIZE
is not set.

Change-Id: Ifa866ad2d0fceac56cbcab3fcd787b8baa55b0f7
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/131115
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>