pw_framebuffer: Moved read/write methods into other classes
pw::framebuffer::Framebuffer is now only a class to manage
ownership (i.e. ability to read and write) of a pixel buffer.
It keeps methods to determine the framebuffer properties
(e.g. dimensions, row bytes), but all methods to set or
get pixels have been moved to two new classes:
FramebufferReader and FramebufferWriter.

Making this change as applications will be using their own
graphics libraries that directly access the pixel data
which don't use the Framebuffer methods. This change
simplifies the Framebuffer class and also makes it clear
to the developer that FramebufferReader and FramebufferWriter
are not performant implementations.

Change-Id: Idbc112cf750b1b283f4e7e94161034a691da0088
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/132970
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>