[roll third_party/pigweed] pw_stream: SocketServer: allow reusing the port
Previously if the server started listening on port 0 it would not use
SO_REUSEADDR, so if later on the server was restarted on the same
(resolved) port number it would fail to bind. In order to fix we are
moving setsocktopt() call so that it happens for port 0 as well.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/133991

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 0b66b67b2406506..c2df1e4942303b0
Roller-URL: https://ci.chromium.org/b/8786546480777717473
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I79a2bfe6c3f9eb2a2ac3ef910d9f00732b82d76e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/134111
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>