[roll third_party/pigweed] soong: Remove host/vendor properties from defaults
The `vendor_available` property indicates that the module in question
may be linked by a vendor module, while `host_supported` allows a host
variant of the module to be built.

For concrete targets like libraries, these properties are very useful in
that they expand the configurations in which the module can be used.
However, the opposite effect occurs when set as part of `cc_defaults`
since these properties are now inherited by the including module and can
conflict with or restrict choices within that module.

One concrete example is the case of trying to build a vendor module that
incorporates a `cc_defaults` with these properties set. This will result
in a build break because it is not permissible to set both `vendor` and
`vendor_available` to true.

Another example is a module incorporating such a default while also
linking a library that was not built with `host_available`. This results
in a build break, because the parent module has inherited the
`host_available` property from the defaults, but it relies on modules
that are not host available.

Finally, note that even if these properties are not included in the
pigweed `cc_defaults`, the module including those defaults is still free
to set these properties as desired. Since there is no advantage to
setting these properties within pigweed `cc_defaults` and a number of
drawbacks, they should be removed.

Test: Full `m` build in Android environment
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/165270

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 565311e620090fe..3db851cdbd209e9
Roller-URL: https://ci.chromium.org/b/8772765161839988097
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I61df3814d118ef8710b154f38f898f1256eb8022
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/165910
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>