[roll third_party/pigweed] pw_cli: Narrow logic around colors
We would always use colors if PW_USE_COLOR was set to a "true" value, or
if it was not set, or if we're connected to a terminal. Change that
logic to default to using colors if connected to a terminal, and
overriding that default if PW_USE_COLOR is set.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/173232

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 3a9727a30897a60..4a65df27bf9e41d
Roller-URL: https://ci.chromium.org/b/8768763302317713873
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I7e5c3bd4e647764d3bf31a10e2b5d365e3003ef7
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/173290
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>