[roll third_party/pigweed] docs: Start the glossary
Creates a new glossary (//docs/glossary.rst) and starts it off with a
definition for "module". Updates other docs to link to the glossary
definition.

Over the long run we may want a more structured approach. Maybe we
store the glossary definitions in a different format (e.g. JSON) to
make it easier to auto-link to the glossary in the rest of the docs.
For now, we should just start with the simplest solution and only
add complexity when necessary.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/172952

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 4a65df27bf9e41d..7ea3c0814e1f56d
Roller-URL: https://ci.chromium.org/b/8768685172482564513
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I894619384b91850b110bde30fa1a673e6934f4e0
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/173490
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>