[roll third_party/pigweed] pw_emu: mock_emu: start listening before making the port available
There is a race condition in mock_emu because we are making the port
available (writing it to a file for clients to read) before marking
the socket as listening.

If the client gets to run after the port is written to the file but
before the TCP thread has a chance to run we would get this error:

======================================================================
ERROR: test_channel_stream (__main__.TestEmulator)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "../../pw_emu/py/tests/frontend_test.py", line 116, in test_channel_stream
    with self._emu.get_channel_stream('gdb') as _:
  File "/b/s/w/ir/x/w/co/pw_emu/py/pw_emu/frontend.py", line 256, in get_channel_stream
    return self._c().get_channel_stream(name, timeout)
  File "/b/s/w/ir/x/w/co/pw_emu/py/pw_emu/core.py", line 533, in get_channel_stream
    sock.connect((host, port))
ConnectionRefusedError: [Errno 111] Connection refused

The fix is to mark the socket for listening before writing the port to
the file.

Original-Bug: 306155313
Test: stress --cpu 512 and run frontend_test.py 100 times without failures
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/176856

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 2f1d686f9f82655..aba67f9a2a71cbc
Roller-URL: https://ci.chromium.org/b/8766876193791408049
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ib1860385f3797bd0bd03a29d21f5663cab37fff0
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/experimental/+/176872
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>